+++ server/server.go
@@ -22,6 +22,7 @@ import (
 	iclient "github.com/influxdata/kapacitor/influxdb"
 	"github.com/influxdata/kapacitor/services/alert"
 	"github.com/influxdata/kapacitor/services/alerta"
+	"github.com/influxdata/kapacitor/services/alertpost"
 	"github.com/influxdata/kapacitor/services/config"
 	"github.com/influxdata/kapacitor/services/deadman"
 	"github.com/influxdata/kapacitor/services/hipchat"
@@ -187,6 +188,7 @@ func New(c *Config, buildInfo BuildInfo, logService logging.Interface) (*Server,
 	s.appendOpsGenieService()
 	s.appendPagerDutyService()
 	s.appendPushoverService()
+	s.appendAlertPostService()
 	s.appendSMTPService()
 	s.appendTelegramService()
 	s.appendSlackService()
@@ -455,6 +457,18 @@ func (s *Server) appendPushoverService() {
 
 	s.SetDynamicService("pushover", srv)
 	s.AppendService("pushover", srv)
+}
+
+func (s *Server) appendAlertPostService() {
+	c := s.config.AlertPost
+	l := s.LogService.NewLogger("[alertpost] ", log.LstdFlags)
+	srv := alertpost.NewService(c, l)
+
+	s.TaskMaster.AlertPostService = srv
+	s.AlertService.AlertPostService = srv
+
+	s.SetDynamicService("alertpost", srv)
+	s.AppendService("alertpost", srv)
 }
 
 func (s *Server) appendSensuService() {