+++ services/alert/service.go
@@ -16,6 +16,7 @@ import (
 	client "github.com/influxdata/kapacitor/client/v1"
 	"github.com/influxdata/kapacitor/command"
 	"github.com/influxdata/kapacitor/services/alerta"
+	"github.com/influxdata/kapacitor/services/alertpost"
 	"github.com/influxdata/kapacitor/services/hipchat"
 	"github.com/influxdata/kapacitor/services/httpd"
 	"github.com/influxdata/kapacitor/services/opsgenie"
@@ -107,6 +108,9 @@ type Service struct {
 	}
 	PushoverService interface {
 		Handler(pushover.HandlerConfig, *log.Logger) alert.Handler
+	}
+	AlertPostService interface {
+		Handler(alertpost.HandlerConfig, *log.Logger) alert.Handler
 	}
 	SensuService interface {
 		Handler(*log.Logger) alert.Handler