+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#1162](https://github.com/influxdata/kapacitor/pulls/1162): Add Pushover integration.
 - [#1221](https://github.com/influxdata/kapacitor/pull/1221): Add `working_cardinality` stat to each node type that tracks the number of groups per node.
 - [#1211](https://github.com/influxdata/kapacitor/issues/1211): Add StateDuration node.
+- [#117](https://github.com/influxdata/kapacitor/issues/117): Add headers to alert POST requests.
 
 ### Bugfixes